pyproject.toml
[build-system]
requires = ["setuptools==64.0.3", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["hyperfetch"]

[project]
name = "hyperfetch"
version = "1.0.12"
authors = [
  { name="Karoline Sund Wahl", email="karolines.wahl@gmail.com" },
]
description = "HyperFetch. A tool to optimize and fetch hyperparameters for your reinforcement learning application."
readme = "README.md"
requires-python = ">=3.9, <3.11.0"
license = {file = "LICENSE"}
keywords = ["reinforcement learning", "hyperparameters", "replication"]

classifiers = [
      'Development Status :: 3 - Alpha',
      'Intended Audience :: Developers',
      'License :: OSI Approved :: BSD License',
      'Operating System :: Microsoft :: Windows :: Windows 11',
      'Operating System :: POSIX :: Linux',
      'Programming Language :: Python :: 3.8',
      'Programming Language :: Python :: 3.9',
      'Programming Language :: Python :: 3.10',
      'Framework :: FastAPI',
]

dependencies = [
      'python-benedict==0.30.0',
      'codecarbon==2.1.4',
      'torch>=1.13.1',
      'gym==0.21.0',
      'pandas==2.0.1',
      'PyYAML==6.0',
      'motor==3.1.2',
      'numpy==1.24.3',
      'optuna==3.1.1',
      'stable-baselines3==1.8.0',
      'rl-zoo3==1.8.0',
      'python-dotenv==1.0.0',
      'starlette==0.26.1'
]

[project.urls]
"Backend" = "https://github.com/karolisw/hyperFetch"
"Website" = "https://github.com/karolisw/hyperFetch/tree/frontend"
"Source" = "https://github.com/karolisw/hyperFetch"

[project.scripts]
tune = "hyperfetch.tuning:tune_cli"
save = "hyperfetch.tuning:save_cli"
